Add missing predict method for restore
We're not doing any :pre assertions on the instance list bc we only
depend on the snapshot being there.

I could have added the list of all snapshots and done a pre-assertion,
and this might be required if we want to add support for creating
snapshots if one is missing, but I won't go there rn, looks messy.